+++ sample/src/commonMain/kotlin/com/github/panpf/sketch/sample/initials.common.kt
@@ -13,6 +13,7 @@ import kotlinx.coroutines.launch
 fun newSketch(context: PlatformContext): Sketch {
     val appSettings = context.appSettings
     return Sketch.Builder(context).apply {
+        disableComponentDetector()
         components {
             supportComposeResources()