+++ sketch-http-hurl/src/commonTest/kotlin/com/github/panpf/sketch/http/hurl/common/test/internal/HurlStackComponentTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.http.hurl.common.test.internal
+
+class HurlStackComponentTest {
+    // TODO test
+}