+++ sketch-http-ktor2/src/nonJvmCommonTest/kotlin/com/github/panpf/sketch/http/ktor2/nonjvm/test/internal/KtorStackComponentNonJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.http.ktor2.nonjvm.test.internal
+
+class KtorStackComponentNonJvmTest {
+    // TODO test
+}