+++ sketch-http-ktor3/src/jvmCommonTest/kotlin/com/github/panpf/sketch/http/ktor3/jvm/test/internal/KtorStackComponentJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.http.ktor3.jvm.test.internal
+
+class KtorStackComponentJvmTest {
+    // TODO test
+}