+++ sketch-http-ktor3/src/nonJvmCommonTest/kotlin/com/github/panpf/sketch/http/ktor3/nonjvm/test/internal/KtorStackComponentNonJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.http.ktor3.nonjvm.test.internal
+
+class KtorStackComponentNonJvmTest {
+    // TODO test
+}