+++ sketch-http-okhttp/src/commonTest/kotlin/com/github/panpf/sketch/http/okhttp/common/test/internal/OkHttpStackComponentTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.http.okhttp.common.test.internal
+
+class OkHttpStackComponentTest {
+    // TODO test
+}